chore: move all reusable section components to one folder
