--- client/src/components/Sections/index.js
@@ -1,4 +1,0 @@
-import SectionWithButtonAndTextGrid from './SectionWithButtonAndTextGrid'
-import SectionWithImage from './SectionWithImage'
-import TextWithTags from './TextWithTags'
-export { SectionWithButtonAndTextGrid, SectionWithImage, TextWithTags }
