+++ client/src/pages/about/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PageWrapper from '../../components/PageWrapper'
-import Hero from '../../sections/common/Hero'
+import { Hero } from '../../sections/common'
 import Content from '../../sections/about/Content'
 import Team from '../../sections/about/Team'
 import CTA from '../../sections/about/CTA'
