+++ client/src/pages/book.js
@@ -3,7 +3,7 @@ import styled from 'styled-components'
 import { Container, Row, Col } from 'react-bootstrap'
 import { Title, Section, Box, Text, Anchor } from '../components/Core'
 import PageWrapper from '../components/PageWrapper'
-import Hero from '../sections/common/Hero'
+import { Hero } from '../sections/common'
 import { usePreviewSubscription } from '../lib/sanity'
 import { getClient } from '../lib/sanity.server'
 import {
