+++ client/src/pages/karriar/index.js
@@ -1,6 +1,6 @@
 import React from 'react'
 import PageWrapper from '../../components/PageWrapper'
-import Hero from '../../sections/common/Hero'
+import { Hero } from '../../sections/common'
 import Content from '../../sections/karriar/Content'
 import Feature from '../../sections/karriar/Feature'
 import Roles from '../../sections/karriar/Roles'
@@ -54,7 +54,7 @@ const Career = ({ data, preview = false, carousel }) => {
       )}
       <PageWrapper headerDark footerDark>
         {preview && <ExitPreviewLink />}
-        <Hero content={hero}/>
+        <Hero content={hero} />
         <Content content={section} carousel={carousel} />
         <Feature content={textGrid} />
         <Roles content={openings} openPositions={data.openPositions} />
