+++ client/src/pages/mvp.js
@@ -7,14 +7,14 @@ import {
   SectionWithButtonAndTextGrid,
   SectionWithImage,
   TextWithTags,
-} from '../components/Sections'
+} from '../sections/common'
 import { usePreviewSubscription } from '../lib/sanity'
 import { getClient } from '../lib/sanity.server'
-import Hero from '../sections/common/Hero'
+import { Hero } from '../sections/common'
 import Content1 from '../sections/mvp/Content1'
 import CTA from '../sections/mvp/CTA'
 import SectionWithImageAndCta from '../sections/mvp/SectionWithImageAndCta'
-import Testimonial from '../sections/startpage/Testimonial'
+import { Testimonial } from '../sections/common'
 import { filterDataToSingleItem, urlFor } from '../utils/helpers'
 
 const Mvp = ({ data, preview = false }) => {
@@ -24,61 +24,57 @@ const Mvp = ({ data, preview = false }) => {
   })
   const mvpPage = filterDataToSingleItem(previewData, preview)
   return (
-    <>
-      <PageWrapper>
-        {preview && <ExitPreviewLink />}
-        {mvpPage.metaTags && (
-          <NextSeo
-            title={mvpPage.metaTags.title}
-            titleTemplate="%s | Iteam"
-            description={mvpPage.metaTags.description}
-            image={urlFor(mvpPage.metaTags.imageWithAlt.asset._ref)}
-            openGraph={{
-              title: mvpPage.metaTags.title,
-              description: mvpPage.metaTags.description,
-              images: [
-                {
-                  url: urlFor(mvpPage.metaTags.imageWithAlt.asset._ref),
-                },
-              ],
-              site_name: 'Iteam',
-            }}
-            twitter={{
-              title: mvpPage.metaTags.title,
-              description: mvpPage.metaTags.description,
-              image: urlFor(mvpPage.metaTags.imageWithAlt.asset._ref),
-              handle: '@iteam1337',
-              site: '@iteam1337',
-              cardType: 'summary_large_image',
-            }}
-          />
-        )}
-        <Hero content={mvpPage?.hero && mvpPage.hero} />
-        {mvpPage?.sectionWithButtonAndTextGrid && (
-          <SectionWithButtonAndTextGrid
-            content={mvpPage.sectionWithButtonAndTextGrid}
-          />
-        )}
-        {mvpPage?.sectionWithImageAndButton && (
-          <Content1 content={mvpPage.sectionWithImageAndButton} />
-        )}
-        {mvpPage?.textWithTags && (
-          <TextWithTags content={mvpPage.textWithTags} />
-        )}
-        {mvpPage?.sectionWithImageAndButton2 && (
-          <SectionWithImageAndCta
-            bg="dark"
-            referenceTo="#book"
-            content={mvpPage.sectionWithImageAndButton2}
-          />
-        )}
-        {mvpPage?.sectionWithImage && (
-          <SectionWithImage content={mvpPage.sectionWithImage} />
-        )}
-        <Testimonial content={data.carousel} />
-        <CTA />
-      </PageWrapper>
-    </>
+    <PageWrapper>
+      {preview && <ExitPreviewLink />}
+      {mvpPage.metaTags && (
+        <NextSeo
+          title={mvpPage.metaTags.title}
+          titleTemplate="%s | Iteam"
+          description={mvpPage.metaTags.description}
+          image={urlFor(mvpPage.metaTags.imageWithAlt.asset._ref)}
+          openGraph={{
+            title: mvpPage.metaTags.title,
+            description: mvpPage.metaTags.description,
+            images: [
+              {
+                url: urlFor(mvpPage.metaTags.imageWithAlt.asset._ref),
+              },
+            ],
+            site_name: 'Iteam',
+          }}
+          twitter={{
+            title: mvpPage.metaTags.title,
+            description: mvpPage.metaTags.description,
+            image: urlFor(mvpPage.metaTags.imageWithAlt.asset._ref),
+            handle: '@iteam1337',
+            site: '@iteam1337',
+            cardType: 'summary_large_image',
+          }}
+        />
+      )}
+      <Hero content={mvpPage?.hero && mvpPage.hero} />
+      {mvpPage?.sectionWithButtonAndTextGrid && (
+        <SectionWithButtonAndTextGrid
+          content={mvpPage.sectionWithButtonAndTextGrid}
+        />
+      )}
+      {mvpPage?.sectionWithImageAndButton && (
+        <Content1 content={mvpPage.sectionWithImageAndButton} />
+      )}
+      {mvpPage?.textWithTags && <TextWithTags content={mvpPage.textWithTags} />}
+      {mvpPage?.sectionWithImageAndButton2 && (
+        <SectionWithImageAndCta
+          bg="dark"
+          referenceTo="#book"
+          content={mvpPage.sectionWithImageAndButton2}
+        />
+      )}
+      {mvpPage?.sectionWithImage && (
+        <SectionWithImage content={mvpPage.sectionWithImage} />
+      )}
+      <Testimonial content={data.carousel} />
+      <CTA />
+    </PageWrapper>
   )
 }
