+++ client/src/sections/common/index.js
@@ -1,0 +1,12 @@
+import SectionWithButtonAndTextGrid from './SectionWithButtonAndTextGrid'
+import SectionWithImage from './SectionWithImage'
+import TextWithTags from './TextWithTags'
+import Testimonial from './Testimonial'
+import Hero from './Hero'
+export {
+  SectionWithButtonAndTextGrid,
+  SectionWithImage,
+  TextWithTags,
+  Testimonial,
+  Hero,
+}
